+++ egui/src/widgets/button.rs
@@ -228,6 +228,7 @@ impl<'a> Widget for Checkbox<'a> {
         desired_size = desired_size.at_least(spacing.interact_size);
         desired_size.y = desired_size.y.max(icon_width);
         let (rect, mut response) = ui.allocate_exact_size(desired_size, Sense::click());
+
         if response.clicked() {
             *checked = !*checked;
             response.mark_changed();