+++ egui/src/widgets/selected_label.rs
@@ -63,7 +63,7 @@ impl Widget for SelectableLabel {
 
         let visuals = ui.style().interact_selectable(&response, selected);
 
-        if selected || response.hovered() {
+        if selected || response.hovered() || response.has_kb_focus() {
             let rect = rect.expand(visuals.expansion);
 
             let corner_radius = 2.0;