+++ CHANGELOG.md
@@ -9,6 +9,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Added ⭐
+
+* You can now give focus to any clickable widget with tab/shift-tab.
+  * Use space or enter to click it.
+  * Use arrow keys to adjust sliders and `DragValue`s.
+
 ### Fixed 🐛
 
 * Fixed secondary-click to open a menu